Changes for lab tests

+++ software/smc/src/adf4351.c
@@ -204,7 +204,7 @@ uint8_t adf4351_set_frequency(uint64_t ullFrequency)
     if(usPhase >= usMOD)
         usPhase = usMOD - 1;
 
-    ADF4351_FREQ = (uint64_t)(ADF4351_PFD_FREQ * (usINT + (float)usFRAC / usMOD)) >> ubRFDivider;
+    ADF4351_FREQ = (uint64_t)(ADF4351_PFD_FREQ * (usINT + (double)usFRAC / usMOD)) >> ubRFDivider;
 
     if(!usFRAC)
         adf4351_rmw_register(2, ~ADF4351_R2_LDF_INT, ADF4351_R2_LDF_INT);
